// apps/desktop/src/lib/generated/types/PresenceStateId.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PresenceStateId = "playing" | "mainMenu" | "hideout" | "heroSelect" | "findingMatch" | "matchFound" | "preGame" | "inMatch" | "streetBrawlRound" | "paused" | "spectating" | "postGame" | "privateLobby" | "practice";