// apps/desktop/src/lib/generated/types/PresenceVariantId.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PresenceVariantId = "solo" | "party" | "unranked" | "ranked" | "heroLabs" | "bots" | "tutorial" | "normal" | "streetBrawl" | "sandbox" | "exploreNyc" | "won" | "lost" | "unscored";